Refactor PKI pkg for better flexibility
