+++ pkg/pki/x509.go
@@ -1,10 +1,7 @@
 package pki
 
 import (
-	"bytes"
 	"crypto/rand"
-	"crypto/rsa"
-	"crypto/tls"
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"encoding/pem"
@@ -12,78 +9,7 @@ import (
 	"fmt"
 	"math/big"
 	"time"
-
-	corev1 "k8s.io/api/core/v1"
 )
-
-var (
-	CACertKey  = "ca.crt"
-	TLSCertKey = "tls.crt"
-	TLSKeyKey  = "tls.key"
-
-	defaultCAValidityDuration   = 4 * 365 * 24 * time.Hour
-	defaultCertValidityDuration = 365 * 24 * time.Hour
-)
-
-const certificatePEMBlockType string = "CERTIFICATE"
-
-type KeyPair struct {
-	Cert    *x509.Certificate
-	Key     *rsa.PrivateKey
-	CertPEM []byte
-	KeyPEM  []byte
-}
-
-func (k *KeyPair) IsValid() bool {
-	return k.Cert != nil && k.Key != nil && len(k.CertPEM) > 0 && len(k.KeyPEM) > 0
-}
-
-func (k *KeyPair) FillTLSSecret(secret *corev1.Secret) {
-	if secret.Data == nil {
-		secret.Data = make(map[string][]byte)
-	}
-	secret.Data[TLSCertKey] = k.CertPEM
-	secret.Data[TLSKeyKey] = k.KeyPEM
-}
-
-func KeyPairFromTLSSecret(secret *corev1.Secret) (*KeyPair, error) {
-	if secret.Data == nil {
-		return nil, errors.New("TLS Secret is empty")
-	}
-	certPEM := secret.Data[TLSCertKey]
-	keyPEM := secret.Data[TLSKeyKey]
-	return KeyPairFromPEM(certPEM, keyPEM)
-}
-
-func KeyPairFromPEM(certPEM, keyPEM []byte) (*KeyPair, error) {
-	if len(certPEM) == 0 || len(keyPEM) == 0 {
-		return nil, errors.New("TLS Secret is empty")
-	}
-	pemBlockCert, _ := pem.Decode(certPEM)
-	if pemBlockCert == nil {
-		return nil, errors.New("Bad certificate")
-	}
-	cert, err := x509.ParseCertificate(pemBlockCert.Bytes)
-	if err != nil {
-		return nil, fmt.Errorf("Error parsing x509 certificate: %v", err)
-	}
-
-	pemBlockKey, _ := pem.Decode(keyPEM)
-	if pemBlockKey == nil {
-		return nil, fmt.Errorf("Bad private key")
-	}
-	key, err := x509.ParsePKCS1PrivateKey(pemBlockKey.Bytes)
-	if err != nil {
-		return nil, fmt.Errorf("Error parsing PKCS1 private key: %v", err)
-	}
-
-	return &KeyPair{
-		Cert:    cert,
-		Key:     key,
-		CertPEM: certPEM,
-		KeyPEM:  keyPEM,
-	}, nil
-}
 
 type X509Opts struct {
 	CommonName   string
@@ -135,8 +61,14 @@ func CreateCA(x509Opts ...X509Opt) (*KeyPair, error) {
 	for _, setOpt := range x509Opts {
 		setOpt(&opts)
 	}
+
+	serialNumber, err := getSerialNumber()
+	if err != nil {
+		return nil, err
+	}
+
 	tpl := &x509.Certificate{
-		SerialNumber: big.NewInt(0),
+		SerialNumber: serialNumber,
 		Subject: pkix.Name{
 			CommonName:   opts.CommonName,
 			Organization: []string{opts.Organization},
@@ -165,8 +97,13 @@ func CreateCert(caKeyPair *KeyPair, x509Opts ...X509Opt) (*KeyPair, error) {
 		return nil, errors.New("CommonName and DNSNames are mandatory")
 	}
 
+	serialNumber, err := getSerialNumber()
+	if err != nil {
+		return nil, err
+	}
+
 	tpl := &x509.Certificate{
-		SerialNumber: big.NewInt(1),
+		SerialNumber: serialNumber,
 		Subject: pkix.Name{
 			CommonName: opts.CommonName,
 		},
@@ -200,7 +137,7 @@ func ParseCertificates(bytes []byte) ([]*x509.Certificate, error) {
 		if block == nil {
 			return nil, errors.New("invalid PEM block")
 		}
-		if block.Type != certificatePEMBlockType {
+		if block.Type != pemBlockTypeCertificate {
 			return nil, fmt.Errorf("invalid PEM certificate block, got block type: %v", block.Type)
 		}
 
@@ -217,22 +154,22 @@ func ParseCertificates(bytes []byte) ([]*x509.Certificate, error) {
 	return certs, nil
 }
 
-func ValidCert(caCert *x509.Certificate, certKeyPair *KeyPair, dnsName string, at time.Time) (bool, error) {
-	if !certKeyPair.IsValid() {
-		return false, errors.New("Invalid certificate KeyPair")
-	}
-	_, err := tls.X509KeyPair(certKeyPair.CertPEM, certKeyPair.KeyPEM)
-	if err != nil {
-		return false, err
-	}
-	parsedCert, err := ParseCertificate(certKeyPair.CertPEM)
-	if err != nil {
-		return false, err
+func ValidCert(caCerts []*x509.Certificate, certKeyPair *KeyPair, dnsName string, at time.Time) (bool, error) {
+	if err := certKeyPair.Validate(); err != nil {
+		return false, fmt.Errorf("invalid keypair: %v", err)
 	}
 
+	certs, err := certKeyPair.Certificates()
+	if err != nil {
+		return false, fmt.Errorf("error getting certificate: %v", err)
+	}
+	cert := certs[0] // leaf certificates should only have a single certificate, not a bundle
+
 	pool := x509.NewCertPool()
-	pool.AddCert(caCert)
-	_, err = parsedCert.Verify(x509.VerifyOptions{
+	for _, cert := range caCerts {
+		pool.AddCert(cert)
+	}
+	_, err = cert.Verify(x509.VerifyOptions{
 		DNSName:     dnsName,
 		Roots:       pool,
 		CurrentTime: at,
@@ -244,50 +181,19 @@ func ValidCert(caCert *x509.Certificate, certKeyPair *KeyPair, dnsName string, a
 }
 
 func ValidCACert(keyPair *KeyPair, dnsName string, at time.Time) (bool, error) {
-	return ValidCert(keyPair.Cert, keyPair, dnsName, at)
+	certs, err := keyPair.Certificates()
+	if err != nil {
+		return false, fmt.Errorf("error getting certificates: %v", err)
+	}
+	return ValidCert(certs, keyPair, dnsName, at)
 }
 
-func createKeyPair(tpl *x509.Certificate, caKeyPair *KeyPair) (*KeyPair, error) {
-	key, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		return nil, err
-	}
-	parent := tpl
-	privateKey := key
-	if caKeyPair != nil {
-		parent = caKeyPair.Cert
-		privateKey = caKeyPair.Key
-	}
+var serialNumberLimit = new(big.Int).Lsh(big.NewInt(1), 128)
 
-	certBytes, err := x509.CreateCertificate(rand.Reader, tpl, parent, key.Public(), privateKey)
+func getSerialNumber() (*big.Int, error) {
+	serialNumber, err := rand.Int(rand.Reader, serialNumberLimit)
 	if err != nil {
 		return nil, err
 	}
-	cert, err := x509.ParseCertificate(certBytes)
-	if err != nil {
-		return nil, err
-	}
-	certPEM, keyPEM, err := pemEncodeKeyPair(certBytes, key)
-	if err != nil {
-		return nil, err
-	}
-
-	return &KeyPair{
-		Cert:    cert,
-		Key:     key,
-		CertPEM: certPEM,
-		KeyPEM:  keyPEM,
-	}, nil
-}
-
-func pemEncodeKeyPair(certificateDER []byte, key *rsa.PrivateKey) (certPEM []byte, keyPEM []byte, err error) {
-	certBuf := &bytes.Buffer{}
-	if err := pem.Encode(certBuf, &pem.Block{Type: "CERTIFICATE", Bytes: certificateDER}); err != nil {
-		return nil, nil, err
-	}
-	keyBuf := &bytes.Buffer{}
-	if err := pem.Encode(keyBuf, &pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(key)}); err != nil {
-		return nil, nil, err
-	}
-	return certBuf.Bytes(), keyBuf.Bytes(), nil
+	return serialNumber, nil
 }
